requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==5.1.8
django-celery-beat==2.7.0
django-configurations==2.5.1
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-import-export==4.3.7
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
drf-yasg==1.21.10
honcho==2.0.0
inflection==0.5.1
kombu==5.5.2
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
